most of my main service are created...sort of

+++ finalProj/src/main/java/finalProj/service/finance/baseServiceInterfaces/ExpenseService.java
@@ -2,5 +2,5 @@ package finalProj.service.finance.baseServiceInterfaces;
 
 import finalProj.domain.finance.Expense;
 
-public interface ExpenseService extends BaseService<Expense, Long> {
+public interface ExpenseService extends BaseService<Expense, Integer> {
 }
